fix: codecheck for tls

Created-by: MinamiK
Commit-by: Cangjie-ChenQian
Merged-by: ChaosJohn
Description: # 仓颉stdx仓PR信息


## 变更内容**（必填）**

修复codecheck 告警


## 变更类型**（必填）**

请描述本次Pull Request变更类型（原因），**请在对应类型的括号内填写Y**：

- 新增需求（ ）
- 问题修复（Y）
- 构建过程或辅助工具变动（）
- 文档更新（）

## 变更内容自检

### 编译器及标准库编译通过截图证明（如涉及新增需求、问题修复、构建过程变动需提供）

请分别提供此Pull Request在你本地环境中的编译器及stdx库编译成功截图。

### 测试用例本地自验证通过截图证明（如涉及新增需求、问题修复需提供）

请提供此Pull Request相关的测试用例在你本地环境中的自验证通过截图。

### 其他信息

在这里可以添加任何与这个 Pull Request 相关的其他说明。


See merge request: Cangjie/cangjie_stdx!157

+++ src/stdx/net/tls/native/tls-impl.c
@@ -578,7 +578,7 @@ static int SslShutdown(SSL* ssl, ExceptionData* exception, DynMsg* dynMsg)
     }
 
     if (result == 0) {
-        int state = DYN_SSL_get_shutdown(ssl, dynMsg);
+        unsigned int state = (unsigned int)DYN_SSL_get_shutdown(ssl, dynMsg);
         if ((state & SSL_SENT_SHUTDOWN) == 0) {
             return CJTLS_NEED_WRITE;
         }
